overlays: move code to overlay_utils.cpp
